Log converter rework
