+++ Gobchat.LogConverter/Gobchat.LogConverter.csproj
@@ -56,6 +56,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Logs\ACT\ACTv1.cs" />
     <Compile Include="External\ChatChannel.cs" />
     <Compile Include="External\Extensions.cs" />
     <Compile Include="External\FFXIVChatChannel.cs" />
@@ -70,6 +71,7 @@
     <Compile Include="LogConverterForm.Designer.cs">
       <DependentUpon>LogConverterForm.cs</DependentUpon>
     </Compile>
+    <Compile Include="Logs\FCL\FCLv1.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <EmbeddedResource Include="LogConverterForm.resx">
